S_scan_ident: Add some comments, white space, braces
This function is complicated, without enough documentation for me to
understand the subtleties; I only studied it enough to change things I
needed to, or which became obvious to me in the process.  Other things
remain undocumented by this commit.

Some of the white space gives improper indentation which will fit a
future commit.

This commit also remove redundant parentheses in one statement