makedef.pl: Eliminate need for embed.fnc sync
makedef.pl has to manually be kept in sync with various other pieces of
code.  It has to know the conditionals that various symbols are compiled
under.  This is extra work, and error prone.  This commit eliminates
that need for embed.fnc syncing.

The relatively new HeaderParser already returns the information about
what conditionals are in effect that we need in its object for each
entry.  This commit just takes advantage of that, and removes the many
lines that were previously needed for this purpose.

Running this under Linux with PLATFORM=test revealed about a dozen
symbols that would have been needlessly exported, such as
Perl_grok_bslash_x(), which is only defined in certain files, and not
globally.