S_scan_ident: Char in middle needt not be IDStart
This fixes a bug in this function, in which it required the second
character in an identifier to be IDStart, instead of IDCont.  This
hasn't been caught because most identifiers are ASCII, and generally for
the purposes of this function in the ASCII range, all \w characters can
be IDStart.