S_scan_ident: Add a check-only option
There is a bug here in which this function is called from S_intuit_more
just to see if there is an identifier in the string it is looking at.
But that call can have "subtle implications on parsing" (according to
the long-standing comments in it).  We need a way to call scan_ident
without side-effects.

This commit adds that capability.  The next will use it.