ParseXS: refactor: remove parser stack
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

The previous commits have changed how the parser tracks file-scoped
  #if / #endif  sections - this is now simply part of the structure
of the AST.

So this commit removes the old parser stack, @{$pxs->{XS_parse_stack}},
which used to be used to track this info. The stack was still present,
but was no longer used to deduce #if state. It also removes the
analyze_preprocessor_statements() method and the test file which used to
test it, t/111-analyze_preprocessor_statements.t.

One minor regression in this commit is that a warning about about
unbalanced #if/#endif within an XSUB or BOOT block used to sometimes
emit an extra hint about what the problem might be:

  print STDERR "    (precede it with a blank line if the matching #if is outside the function)\n"
      if $self->{XS_parse_stack}->[-1]{type} eq 'if';

Since it's no longer trivial to check whether we're currently within a
global #if/endif section, I've kept the 'unbalanced' warning, but
deleted the additional hint.

If anybody wants to add it back, you'll need to add a field to the
EU::PXS class such as

    $pxs->{current_cpp_scope}

which keeps track of which is the current (innermost) Node::cpp_scope
node being parsed. Then check_conditional_preprocessor_statements() can
use it in place of the "$self->{XS_parse_stack}->[-1]{type} eq 'if';"
test.