Create BYTES_REMAINING_IN_WORD()
This macro encapsulates the task of finding how far until the next word
boundary the passed-in address is.

There are several places that could use this, but instead of converting
use this in those places, the next commit will create macros that depend
on this one and those places will instead convert to use those other new
macros.