regen/embed.pl: Avoid use of hard-coded list
The list consists of exactly the functions that have the O flag set in
embed.fnc.  No need to keep this data twice.  The entries are trivially
generatable from existing entries as we go along

And those generated entries have the added advantage of not using the
short name, so potentially less name space pollution