Create WORTH_PER_WORD_LOOP(), WORTH_PER_WORD_LOOP_BINMODE()
There are several places in the perl core that, for performance,  use
word-at-a-time operations on byte data when the data to be processed is
long enough to overcome the extra setup overhead required.

The code that does this is not immediately obvious, and is currently
repeated at each such place.

This macro creates two macros that encapsulate this logic, making each
place that uses them easier to read.

One macro is for data that isn't dependent on the character set.  The
other is for character data.  EBCDIC data is not suitable for per-word
operation, so the this macro always returns false on an EBCDIC platform.
This allows for the removal of some EBCDIC #ifdefs in our code base.