gv_magicalize: Quickly rule non-magical input out
This uses the data structure introduced in the previous commit to
quickly test the input first character.  If it isn't a potential magical
one, it could apply to CORE, so move the block that checks for that to
here, eliminating a conditional.  In either case, no need to look further.