gv_magicalize: Refactor
This refactors to eliminate redundant code.  Some things are magical
only if we are using the main stash; others in any stash; one only in
PL_debstash.

Previously, the switches were structured thusly:
 1) if we aren't using the main stash, handle things not requiring the
        main stash
 2) if we are using the main stash, handle all len > 1 things that can
        be in the main stash.  This duplicates much of item 1)
 3) if we are using the main stash, handle all len == 1 things that can
        be in the main stash.  This duplicates some of item 1)

The new structure is
 if (len > 1) {
    1) handle len > 1 things not requiring main stash, regardless of the
       stash we are in
    2) handle len > 1 things requiring main stash
 } else {
    3) handle len == 1 things, regardless of the stash we are in.
 }

This removes the duplicated code.

The case for 'a' and 'b' are special.  When 'a' stands for "args" it is
len > 1 and that is handled in 1).

But 'a' can also mean a single character, as 'b' always does.  These
cases are handled in 3).  These are the only two len == 1 characters
that don't have to be in the main package, so there is an extra
conditional clause to allow that.