ParseXS: refactor: remove $ExtUtils::ParseXS::END
(This commit is part of a series which will extend the AST parse tree
from just representing individual XSUBs to representing the whole XS
file.)

The parser, when parsing an XSUB (but not any other parts of an XS file),
pushes a special token onto the end of the list of lines in
@{$pxs->{line}} which make up the current paragraph. This token,

    $ExtUtils::ParseXS::END = "!End!\n\n";

(along with a trailing ':') is designed  to look like an impossible
keyword which can't actually appear in the source code (due to the
multiple newlines).

It looks like it was originally added in perl5.002 to make the parsing
code easier, but I don't really understand why. It just makes the
parser harder to understand.

So this commit removes it, and just relies on @{$pxs->{line}} being zero
to detect the end of the paragraph. This change doesn't alter the C code
generated from any of the XS files bundled with perl.