regcomp.c: Need to account for UTF group name
I found this by reading the code.  Prior to this commit, the parse
pointer was advanced by one byte; it should be advanced by one
character.  As long as the the character was ASCII, things worked.

I looked through the regcomp.c source for other mis-use of the macro
changed by this commit; none were obvious.