ParseXS: refactor: fetch_para(): remove POD loop
Remove the loop which says 'foreach contiguous section of pod, delete
it'. Replace it with with something which removes just *one* section of
POD, then relies on the main loop to process the next line, which may
well be the start of another POD section.

Fewer nested loops is less cognitive load.

Should be no functional changes.