ParseXS: process line continuations after POD
Within fetch_para(), the at-end-of-POD code and the main loop both had
separate (but similar) chunks of code to read in the next line, with the
latter one being more complete. In particular, the latter had logic to
process '\' continuation lines, while the POD one didn't. This meant
that this gave a parse error:

    =pod
    =cut
    void foo(int i, \
             int j)

The fix is to make them both use the same code to read the next line.