ParseXS: process line continuations after TYPEMAP
(This is similar to the previous commit, which fixed the same bug for
POD)

Within fetch_para(), the at-end-of-TYPEMAP code and the main loop both
had separate (but similar) chunks of code to read in the next line, with
the latter one being more complete. In particular, the latter had logic
to process '\' continuation lines, while the TYPEMAP one didn't. This
meant that this gave a parse error:

    TYPEMAP: <<EOF
       ...
    EOF
    void foo(int i, \
             int j)

The fix is to make them both use the same code to read the next line.
This bug was probably actually introduced by me during the refactoring
in this branch.

By using the same read-next-line code, it also means that an immediately
following blank line is now stripped of whitespace. I'm not sure whether
that actually affects anything in the real word. Also, any trailing
lines in typemap are now stripped. Again, this shouldn't affect
anything.