ParseXS: refactor: fetch_para(): simplify logic
Simplify the logic of the code which handles C-preprocessor directives
where there's not the (usually required) blank line directly before and
after an XSUB; e.g.

    #if X
    int foo(...)
       ...
    #endif

This commit splits the logic into two halves: the first is responsible
for handling conditionals directly *before* the XSUB, and the other half
for ones during and directly after the XSUB. Then the second half is
further simplified.

I think this makes the code easier to understand.

I *think* the new code is logically equivalent.