intuit_more: Remove over allocation of buffer
I don't know why the code here uses a buffer 4 times bigger than what is
the maximum an identifier can be.  One might think this is because
Unicode characters can be 4 bytes long, but this number has been in
effect since before Perl knew about Unicode; or maybe it was to avoid
a potential buffer overflow.

But in any event, I'm pretty certain there is no need for that now.  The
buffer is now allocated to handle the maximum Unicode-needed size, and
buffer overflow is checked for and avoided.