Collect names of multiple erroring parameters
Rather than stop immediately at the first failure, we can be a bit more
helpful by collecting up potentially multiple names at once.

We limit the list to 5 elements, just to limit the potential output in
the case of accidentally passing a list of hundreds or thousands of
elements into such a function.