[PATCH] Docu suggestion for named parameters in signatures
Add a paragraph that callers need to specify all positional parameters
before named ones, even if the positional parameters have defaults.

Delete a sentence stating that there is no way for the caller to provide
a value for a named parameter after an optional positional parameter.
There _is_ a way, you just need to provide values for positional
parameters.