grok_bin_oct_hex: Use upper bound, not length remaining
Creating an upper limit to parse allows us to write while (s < e) for
example, and that limit is constant, requiring fewer operations than the
other way, where the remaining length keeps getting changed.

It also allows this commit to move an 's++' a couple of lines to get rid
of comparing against the number '8' which could get out of sync.