Avoid redefining SvREADONLY_on in gv.c as that causes confusion
When reading the previous code, it was confusing to see calls to
`SvREADONLY_on`, as without being aware the macro was redefined just
above this function, the reader is unlikely to be aware that here it
doesn't set the SVf_PROTECT flag.

We should instead define a custom-purpose macro here and use that, to
make its operation much clearer.