grok_bin_oct_hex: fix broken return flags
Apparently no one has tried to use these before.  These flags are to
suppress the display of certain warnings, but to instead return that the
suppression happened in output flags.  The output flags were not getting
set.

I'm not adding a separate test, because a future commit will cause this
feature to be used regularly.