scan_num: Macroize common code
This avoids repeating code snippets.  It also changes things so adjacent
underscores are all absorbed at once (and warned about).  That means we
no longer have to keep track of if the previous character was an
underscore, so the variable that did that is removed.

Only two checks need be done for running off either end of the buffer.
The buffer is NUL-terminated, so if we see an underscore in the current
position, the next position exists (there is a NUL there if nothing
else); and the macro that looks behind one position is called in only
one place where we haven't always parsed beyond the first character.