regmatch(): improve savestack ix debugging output
Under

    use re Debug => 'STATE';

S_regmatch() produces various bits of debugging output regarding
savestack manipulation; e.g. when the current index is saved; or when
the savestack is popped back to that index on success or failure etc.

Unfortunately that output is patchy and inconsistent. For example it
often talks confusingly about "Setting an EVAL scope" even when an EVAL
hasn't been executed. And it doesn't log all savestack manipulation.

After this commit there is a standard set of debugging messages, such
as:

    savestack: snapping ix=14
    savestack: regcppush filled in range [14..18)
    savestack: regcppop freed in range [14..18)
    savestack: freeing in range [10..14)

which are consistently applied.

There should be no change in functionality unless that specific
debugging is enabled.