Don't treat class attributes as sub attributes in perly.y
We need to use a dedicated grammar type; or at the very least, not use
`subattrlist` for doing this. The previous use of `subattrlist` had the
side-effect of passing every attribute list for a class via the
`apply_builtin_cv_attributes()` function. That was pointlessly
redundant, though currently harmless. It may stop being harmless in a
future improvement to the way attributes are handled, so best not to do
that any more.

The new structure repurposes the previous `myattrlist` to be a
non-optional attribute list, and defines `optattrlist` to use for both
class and field attributes.