Clear up ambiguous use of 'try block'
Way back in 2001 several tests were added to t/op/goto.t which
referred to a concept called a 'try block':

    commit 9c5794fe8bd5803990d11472f5ebcdba0e7877cc
    Author:     Robin Houston <robin@cpan.org>
    AuthorDate: Wed Mar 14 01:52:51 2001 +0000
    Commit:     Jarkko Hietaniemi <jhi@iki.fi>
    CommitDate: Wed Mar 14 02:41:54 2001 +0000
    ...
    diff --git a/t/op/goto.t b/t/op/goto.t
    index 246184c56b..6f9e3493d5 100755
    --- a/t/op/goto.t
    +++ b/t/op/goto.t
    ...
    +# Does goto work correctly within a try block?

That language is still in use in two files -- but since we now have
a 'try' keyword and a feature 'try', such language is now ambiguous.

Here we change 'try block' to 'eval block' in those test descriptions
and comments where we're clearly speaking about 'eval BLOCK'.