Configure: Allow LC_ALL syntax override
Fixes #22992

The problem here had to do with cross-compilation.  There was no way to
override Perl's calculation of how LC_ALL is represented.  I had not put
one in because it is hard to get it right manually.  But this doesn't
work in cross-compilation; it's better to have a difficult-to-get-right
way than no way at all.  Porting/Glossary is updated to have an example.